more cleanup all over the place
git-svn-id: svn+ssh://gate.spectrum.uni-bielefeld.de/homes/tbaumann/inpro-svnrepo/trunk@2801 19460358-522e-0410-942e-fdb976ae3ca2

+++ src/inpro/features/TimeShiftingAnalysis.java
@@ -142,7 +142,7 @@ public class TimeShiftingAnalysis implements Resetable {
 			// while we're there, also compute some other measures
 			mean = (sumX / n);
 			stddev = Math.sqrt((sumXX / n) - mean * mean);
-			meanDifference = accumulatedDifferences / ((double) (n - 1));
+			meanDifference = accumulatedDifferences / (n - 1.0);
 			mse = ((sumSqDevX) - (sumSqDevT) * (slope * slope)) / (n*n);
 			range = (maxDP != null) ? (maxDP.x - minDP.x) : 0;
 			// int timeSpanned = dataPoints.getFirst().t - dataPoints.getLast().t;
